fix signal names
Signal IDs are prefixed by a byte. This switches between the
number ranges "item", "fluid" and "virtual". There are *similar*
prototype types "item" and "fluid" and "virtual-signal" but they
do not match perfectly - even besides the mismatch "virtual" vs.
"virtual-signal":

It seems that the index lists the code 66 for `cliff-explosives`
in `capsule`, not in `item`. But `Ctrl-Shift-E` lists the code 66
in the section `item`. So it seems to be some subtype thing.

Which is FUUBAR because the subtypes are not deducible from the
dat-file.

To make things more "interesting": `cliff-explosives` has another
code 68 in `Ctrl-Shift-E` within `entity.