parse turrets
TL;DR: Turrets are strange.

Turrets have much unparsable data which does not show up in the export file.
This includes -- only for some types of turrrets -- the direction. So the
direction these turrets are constructed does not survive an export/import cylce.

Also the position of artillery turrets seems to be always absolute.