python: adapt python output to better match perl
Perl changes freely between int and float representation. Therefore the
JSON output for numbers omitts a trailing ".0" for floats without a fraction.
Pyhton seems to output ".0" for floats.

Interesting point: The export format seems to align with Perl, not Python.
So the bets are open how this will be resolved.