support item requests for all entities
This is a followup to:

	6af4375c ("cleanup comments for ep_items", 2021-11-03)

The vanilla games supports `items` only for a few entities -- mostly modules in
assemblers, labs, ... But mods (and hacking the blueprint strings) can add
arbitrary item requests to any entity! If an entity cannot take in some
delivered item then the request counter is decremented and the construction
robot returns the items. So if an inserter requests 200 firearm magazines then
the robots will do quire some shutteling.